tests/wpt/tests/html/browsers/history/the-location-interface/location_hostname.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_hostname</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    test(function () {
      var hostname = location.hostname;
      var url = location.href;

      var pos = url.indexOf("//");
      if (pos != -1) {
        url = url.substr(pos+2, url.length-pos-2);
        pos = url.indexOf(":");
        if (pos != -1) {
          url = url.substr(0, pos);
        } else {
          pos = url.indexOf("/");
          if (pos != -1)
            url = url.substr(0, pos);
        }
      }

      assert_equals(hostname, url, "hostname");

    }, "location hostname");
    </script>
  </body>
</html>
